--- src/main/java/kr/hhplus/be/server/domain/coupon/CouponRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.coupon;
-
-public interface CouponRepository {
-
-}
