+++ src/main/java/kr/hhplus/be/server/domain/coupon/IssuedCoupon.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.pointhistory;
+package kr.hhplus.be.server.domain.coupon;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,7 +7,7 @@ import jakarta.persistence.Id;
 import kr.hhplus.be.server.domain.common.Base;
 
 @Entity
-public class PointHistory extends Base {
+public class IssuedCoupon extends Base {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
