+++ src/main/java/kr/hhplus/be/server/domain/coupon/repository/CouponRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.coupon.repository;
+
+public interface CouponRepository {
+
+}
