+++ src/main/java/kr/hhplus/be/server/domain/coupon/repository/IssuedCouponRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.coupon.repository;
+
+public interface IssuedCouponRepository {
+
+}
