+++ src/main/java/kr/hhplus/be/server/domain/order/Order.java
@@ -4,9 +4,11 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import kr.hhplus.be.server.domain.common.Base;
 
 @Entity
+@Table(name = "order_t")
 public class Order extends Base {
 
     @Id
