--- src/main/java/kr/hhplus/be/server/domain/order/OrderRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.order;
-
-public interface OrderRepository {
-
-}
