+++ src/main/java/kr/hhplus/be/server/domain/order/repository/OrderItemRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.order.repository;
+
+public interface OrderItemRepository {
+
+}
