+++ src/main/java/kr/hhplus/be/server/domain/order/repository/OrderRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.order.repository;
+
+public interface OrderRepository {
+
+}
