--- src/main/java/kr/hhplus/be/server/domain/orderitem/OrderItemRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.orderitem;
-
-public interface OrderItemRepository {
-
-}
