--- src/main/java/kr/hhplus/be/server/domain/orderitem/OrderItemService.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.domain.orderitem;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class OrderItemService {
-
-}
