--- src/main/java/kr/hhplus/be/server/domain/payment/PaymentRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.payment;
-
-public interface PaymentRepository {
-
-}
