+++ src/main/java/kr/hhplus/be/server/domain/payment/repository/PaymentRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.payment.repository;
+
+public interface PaymentRepository {
+
+}
