--- src/main/java/kr/hhplus/be/server/domain/point/Point.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.domain.point;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import kr.hhplus.be.server.domain.common.Base;
-
-@Entity
-public class Point extends Base {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-}
