--- src/main/java/kr/hhplus/be/server/domain/point/PointRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.point;
-
-public interface PointRepository {
-
-}
