--- src/main/java/kr/hhplus/be/server/domain/point/PointService.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.domain.point;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class PointService {
-
-}
