--- src/main/java/kr/hhplus/be/server/domain/pointhistory/PointHistoryRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.pointhistory;
-
-public interface PointHistoryRepository {
-
-}
