--- src/main/java/kr/hhplus/be/server/domain/pointhistory/PointHistoryService.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.domain.pointhistory;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class PointHistoryService {
-
-}
