--- src/main/java/kr/hhplus/be/server/domain/product/ProductRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.product;
-
-public interface ProductRepository {
-
-}
