+++ src/main/java/kr/hhplus/be/server/domain/product/repository/ProductRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.product.repository;
+
+public interface ProductRepository {
+
+}
