+++ src/main/java/kr/hhplus/be/server/domain/user/PointHistory.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.usercoupon;
+package kr.hhplus.be.server.domain.user;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,7 +7,7 @@ import jakarta.persistence.Id;
 import kr.hhplus.be.server.domain.common.Base;
 
 @Entity
-public class UserCoupon extends Base {
+public class PointHistory extends Base {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
