--- src/main/java/kr/hhplus/be/server/domain/user/UserRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.user;
-
-public interface UserRepository {
-
-}
