+++ src/main/java/kr/hhplus/be/server/domain/user/repository/PointHistoryRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.user.repository;
+
+public interface PointHistoryRepository {
+
+}
