--- src/main/java/kr/hhplus/be/server/domain/usercoupon/UserCouponRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.usercoupon;
-
-public interface UserCouponRepository {
-
-}
