--- src/main/java/kr/hhplus/be/server/domain/usercoupon/UserCouponService.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.domain.usercoupon;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserCouponService {
-
-}
