+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/coupon/issuedcoupon/IssuedCouponRepositoryImpl.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.infrastructure.jpa.coupon.issuedcoupon;
+
+import kr.hhplus.be.server.domain.coupon.repository.IssuedCouponRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class IssuedCouponRepositoryImpl implements IssuedCouponRepository {
+
+}
