+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/order/OrderRepositoryImpl.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.infrastructure.jpa.order;
 
-import kr.hhplus.be.server.domain.order.OrderRepository;
-import kr.hhplus.be.server.domain.user.UserRepository;
+import kr.hhplus.be.server.domain.order.repository.OrderRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
