+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/order/orderitem/OrderItemJpaRepository.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.infrastructure.jpa.order.orderitem;
+
+import kr.hhplus.be.server.domain.order.OrderItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderItemJpaRepository extends JpaRepository<OrderItem, Long> {
+
+}
