+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/order/orderitem/OrderItemRepositoryImpl.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.infrastructure.jpa.order.orderitem;
+
+import kr.hhplus.be.server.domain.order.repository.OrderItemRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class OrderItemRepositoryImpl implements OrderItemRepository {
+
+}
