--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/orderitem/OrderItemRepositoryImpl.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.orderitem;
-
-import kr.hhplus.be.server.domain.orderitem.OrderItemRepository;
-import kr.hhplus.be.server.domain.user.UserRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class OrderItemRepositoryImpl implements OrderItemRepository {
-
-}
