--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/point/PointJpaRepository.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.point;
-
-import kr.hhplus.be.server.domain.order.Order;
-import kr.hhplus.be.server.domain.point.Point;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PointJpaRepository extends JpaRepository<Point, Long> {
-
-}
