--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/point/PointRepositoryImpl.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.point;
-
-import kr.hhplus.be.server.domain.point.PointRepository;
-import kr.hhplus.be.server.domain.user.UserRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class PointRepositoryImpl implements PointRepository {
-
-}
