--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/pointhistory/PointHistoryJpaRepository.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.pointhistory;
-
-import kr.hhplus.be.server.domain.pointhistory.PointHistory;
-import kr.hhplus.be.server.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PointHistoryJpaRepository extends JpaRepository<PointHistory, Long> {
-
-}
