--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/pointhistory/PointHistoryRepositoryImpl.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.pointhistory;
-
-import kr.hhplus.be.server.domain.pointhistory.PointHistoryRepository;
-import kr.hhplus.be.server.domain.user.UserRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class PointHistoryRepositoryImpl implements PointHistoryRepository {
-
-}
