+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/product/ProductRepositoryImpl.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.infrastructure.jpa.product;
 
-import kr.hhplus.be.server.domain.product.ProductRepository;
-import kr.hhplus.be.server.domain.user.UserRepository;
+import kr.hhplus.be.server.domain.product.repository.ProductRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
