+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/user/pointhistory/PointHistoryJpaRepository.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.infrastructure.jpa.user.pointhistory;
+
+import kr.hhplus.be.server.domain.user.PointHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PointHistoryJpaRepository extends JpaRepository<PointHistory, Long> {
+
+}
