+++ src/main/java/kr/hhplus/be/server/infrastructure/jpa/user/pointhistory/PointHistoryRepositoryImpl.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.infrastructure.jpa.user.pointhistory;
+
+import kr.hhplus.be.server.domain.user.repository.PointHistoryRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class PointHistoryRepositoryImpl implements PointHistoryRepository {
+
+}
