--- src/main/java/kr/hhplus/be/server/infrastructure/jpa/usercoupon/UserCouponRepositoryImpl.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.jpa.usercoupon;
-
-import kr.hhplus.be.server.domain.user.UserRepository;
-import kr.hhplus.be.server.domain.usercoupon.UserCouponRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class UserCouponRepositoryImpl implements UserCouponRepository {
-
-}
